+++ CHANGELOG.md
@@ -11,6 +11,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - Method to check if a string is a valid IP V4 ([#2](https://github.com/khalyomede/ip/issues/2)).
 - Method to check if a string is a valid IP V6 ([#2](https://github.com/khalyomede/ip/issues/2)).
+- Method to parse any string to an IP V4 or V6, e.g. an Address ([#1](https://github.com/khalyomede/ip/issues/1)).
 
 ## [0.2.0] - 2025-10-21